add checking if requester user is channel creator while deleting a channel

+++ apps/server/application/commandHandlers/deleteChannelCommandHandler/DeleteChannelCommandHandlerImpl.cpp
@@ -3,6 +3,7 @@
 #include <format>
 
 #include "../../errors/ResourceNotFoundError.h"
+#include "server/application/errors/OperationNotValid.h"
 
 namespace server::application
 {
@@ -19,6 +20,14 @@ void DeleteChannelCommandHandlerImpl::execute(const DeleteChannelCommandHandlerP
     if (!existingChannel)
     {
         throw errors::ResourceNotFoundError{std::format("Channel with id {} not found.", payload.channelId)};
+    }
+
+    const auto creatorId = existingChannel->get()->getCreatorId();
+
+    if (payload.requesterUserId != creatorId)
+    {
+        throw errors::OperationNotValid{std::format("User with id {} is not creator of the channel with id {}.",
+                                                    payload.requesterUserId, payload.channelId)};
     }
 
     channelRepository->deleteChannel({**existingChannel});
